feat: Add CSV/JSON export APIs and shorter time range options
Export Endpoints:
- Add /api/textfiles/export_csv POST endpoint for filtered CSV export
- Add /api/data/export/<format> GET endpoint for CSV and JSON export
- Include channel names in exported data

Dashboard Time Ranges:
- Add Last 5 Mins, Last 10 Mins, Last 30 Mins options
- Change hours parameter from int to float for sub-hour precision
- Default view changed to Last 1 Hour

+++ data-logger-project/app.py
@@ -296,7 +296,7 @@ def api_latest_data():
 @app.route('/api/data/historical')
 def api_historical_data():
     try:
-        hours = request.args.get('hours', 24, type=int)
+        hours = request.args.get('hours', 24, type=float)
 
         print(f"[DEBUG] /api/data/historical called - Requesting last {hours} hours")
